% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{fetch_new_tweets}
\alias{fetch_new_tweets}
\title{Fetch New Tweets}
\usage{
fetch_new_tweets(con, twitter_handle, ...)
}
\arguments{
\item{con}{db connection}

\item{twitter_handle}{Twitter handle for fetching tweets}

\item{...}{passed to rtweet::get_timeline}
}
\value{
dataframe of latest tweets
}
\description{
Fetch New Tweets
}
\examples{
# NO RUN
}
