% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{filter_retweets}
\alias{filter_retweets}
\title{Filter out retweets from rtweet data}
\usage{
filter_retweets(data)
}
\arguments{
\item{data}{Dataframe}
}
\value{
dataframe
}
\description{
Filter out retweets from rtweet data
}
\examples{
# Not Run
}
