% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{forecast_set}
\alias{forecast_set}
\title{Get Data For Forecasting}
\usage{
forecast_set(con)
}
\arguments{
\item{con}{DB connection}
}
\value{
a dataframe
}
\description{
Get Data For Forecasting
}
\examples{
# Note Run
}
