% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_number}
\alias{get_number}
\title{Extract number from tweet}
\usage{
get_number(text, slug)
}
\arguments{
\item{text}{cleaned text}

\item{slug}{slug to grab number from}
}
\value{
int
}
\description{
Extract number from tweet
}
\examples{
# NOT RUN
}
