% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{observed_set}
\alias{observed_set}
\title{Get Observations From Today}
\usage{
observed_set(con)
}
\arguments{
\item{con}{DB connection}
}
\value{
A dataframe
}
\description{
Get Observations From Today
}
\examples{
# Not Run
}
