% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{prediction_times}
\alias{prediction_times}
\title{Prediction Times For Model Plotting}
\usage{
prediction_times(con)
}
\arguments{
\item{con}{DB connection}
}
\value{
a dataframe of times
}
\description{
Prediction Times For Model Plotting
}
\examples{
# Not Run
}
