% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{wr_numbers}
\alias{wr_numbers}
\title{Extract data for training}
\usage{
wr_numbers(con)
}
\arguments{
\item{con}{Database Connection}
}
\value{
tibble of training data
}
\description{
Extract data for training
}
\examples{
#Not Run
}
