% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{write_to_db}
\alias{write_to_db}
\title{Write a dataframe to database}
\usage{
write_to_db(data, con)
}
\arguments{
\item{data}{dataframe}

\item{con}{db connection}
}
\description{
Write a dataframe to database
}
\examples{
# NORUN
}
